Feature/login with apple (#227)
* Login with apple button if platform is iOs

* Extract social media buttons to separate component

* Login with apple instead of Facebook button for iOs

* Regenerated pod lockfile

---------
