+++ ios/Runner/Info.plist
@@ -68,5 +68,7 @@
         </array>
       </dict>
     </array>
+	<key>UIApplicationSupportsIndirectInputEvents</key>
+	<true/>
 </dict>
 </plist>
